requirements.txt
pandas
numpy
snowflake-connector-python
matplotlib
python-dotenv
sqlalchemy